fix: Improve the image and the anchor elements
The image elements were improved upon by adding the alt attributes and
width and height attributes also. The alt attribute helps screen reading
tools to identify the image in the website. The width and height
attributes resize the food image for better view.

The anchor element was improved by adding the target and rel attribute.
These improved security of the website.